Lessons.Lesson08

Description

Notes taken by Julija Mikeliūnaitė

Documentation

fl :: [Int]

Monad is the most powerful class. Bind method (>>=) gives the illusion of sequential computation.

Functors ignore the order of execution.

| Minimal implementation of a functor: fmap :: Functor f => (a -> b) -> f a -> f b It takes a pure function (which takes a and returns b), a functor of a, and returns a functor of b.

The function was applied to every element of the list.

>>> fl
[5,4]

fm :: Maybe Integer

If the list is empty, the result is also an empty list.

>>> fm
Nothing

fe :: Either String Integer

fe' :: Either Integer Integer

The left value is considered a "bad" value.

>>> fe'
Left 41

fio :: IO String

Takes an input and returns the value with "!"" in the end. Returned type is IO String. IO is a type which presents computations which originate in the external world.

p :: [Integer]

Applicative functor is between functor and monad.

pure is pretty much the same as return: pure :: Applicative f => a -> f a return :: Monad m => a -> m a

>>> p
[5]

am :: Maybe Integer

Signature *, sometimes called "spaceship". (*) :: Applicative f => (a -> b) -> f a -> f b It is similar to fmap, but instead of a Functor, it uses an Applicative.

Signature $ is synonymous with fmap.

>>> am
Just 46

am' :: Maybe Integer

am'' :: Maybe Integer

al :: [Integer]

Sources of values are completely independent, computed in parallel.

>>> al
[2,3,3,4,4,5]

ml :: [Integer]

Sources of values depend on values that are above, line-by-line.

>>> ml
[2,3,3,4,4,5]

newtype Parser a

newtype is used if ADT has a single constructor.

Constructors

Parser

Fields

runParser :: String -> Either String (a, String)

parseLetter :: Parser Char

If a number is parsed, then it returns Left value.

>>> runParser parseLetter "skfhsdk"
Right ('s',"kfhsdk")

threeLetters :: Parser String

parseString :: Parser String

parseNonEmptyString :: Parser String