Lessons.Lesson08

Documentation

fio :: IO String

newtype Parser a

al' and ml are basically the same thing: In al, sources of values are completely independent, computed in parallel. In ml, they depend on values that are above, line-by-line.

Constructors

Parser

Fields

runParser :: String -> Either String (a, String)

threeLetters :: Parser String