Lessons.Lesson11

Documentation

data Expr

Lists are constructed using :, therefore the first n elements can be checked one by one.

Eager calculations - trying to compute all data at once and store everything in memory. It can be achieved using libraries with the keyword Strict.

seq function: seq :: a -> b -> b Simplest explanation: the return value is bottom if a is bottom, otherwise it is b.

Returns 5 and ensures that everything was calculated/checked. >>> seq 4 5 5

Returns 5, because seq removes the top layer of what is uncalculated. >>> seq [error Aj] 5 5

Does all of the calculations, no matter the layer. >>> deepseq ([error Aj]:: [Int]) 5 Aj

DSL - domain specific language In our DSL's we had differend commands - a dictionary of tasks which we can use within our language. A classic example - calculator.

Constructors

Lit Integer

Add Expr Expr

Mul Expr Expr

Neg Expr

eval :: Expr -> Integer

This is used to define how the program should act with different expressions. Evaluate >>> eval prog1 -15